Add new `compiler` syntax and empty `host`
In `conda-build` 3, the new `compiler` syntax is used to specify a
compiler and ensure certain compiler runtime libraries are added as
dependencies. Given this requires a C compiler, go ahead and require
that one.

Also add an empty `host` to split the `build` tool requirements from
where the package will be installed.